+++ server/api/src/builder/user.builder.ts
@@ -1,6 +1,6 @@
 import { BuilderCommon } from './builder';
 import { User } from '../domain/user/user.entity';
-import { DevField } from '../domain/user/dev-field.entity';
+import { DevField } from '../domain/field/dev-field.entity';
 import { Follow } from '../domain/user/follow.entity';
 import { History } from '../domain/history/history.entity';
