--- server/api/src/domain/user/dev-field.entity.ts
@@ -1,14 +1,0 @@
-import { Column, Entity, OneToMany, PrimaryGeneratedColumn } from 'typeorm';
-import { User } from './user.entity';
-
-@Entity()
-export class DevField {
-  @PrimaryGeneratedColumn()
-  id: number;
-
-  @Column({ type: 'varchar', length: 45, unique: false })
-  name: string;
-
-  @OneToMany((type) => User, (user) => user.devField)
-  users: User[];
-}
