--- server/api/src/domain/user/repository/dev-field.repository.ts
@@ -1,9 +1,0 @@
-import { EntityRepository, Repository } from 'typeorm';
-import { DevField } from '../dev-field.entity';
-
-@EntityRepository(DevField)
-export class DevFieldRepository extends Repository<DevField> {
-  async findDevById(devField: number): Promise<DevField> {
-    return await this.findOneOrFail({ where: { id: devField } });
-  }
-}
