+++ server/api/src/domain/user/service/user.service.ts
@@ -2,9 +2,9 @@ import { Injectable } from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
 import { UserException, DevFieldException } from '../../../exception';
 import { User } from '../user.entity';
-import { DevField } from '../dev-field.entity';
+import { DevField } from '../../field/dev-field.entity';
 import { ImageService } from '../../image/service/image.service';
-import { DevFieldRepository } from '../repository/dev-field.repository';
+import { DevFieldRepository } from '../../field/repository/dev-field.repository';
 import { UserRepository } from '../repository/user.repository';
 import { UserUpdateDto } from '../dto/user-update.dto';
 import { UserResponseDto } from 'src/domain/auth/dto/user-response.dto';
